+++ account_reconcile_oca/__init__.py
@@ -1,2 +1,2 @@
-from . import models
+from . import models, wizards
 from .hooks import post_init_hook
