+++ account_reconcile_oca/wizards/account_account_reconcile_data.py
@@ -1,0 +1,13 @@
+# Copyright 2023 Dixmit
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class AccountAccountReconcileData(models.TransientModel):
+    _name = "account.account.reconcile.data"
+    _description = "Reconcile data model to store user info"
+
+    user_id = fields.Many2one("res.users", required=True)
+    reconcile_id = fields.Integer(required=True)
+    data = fields.Serialized()
